MartinSearchDocuments/requirements.txt
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
idna==2.10
msal==1.10.0
Office365-REST-Python-Client==2.3.1
pip==21.0.1
pycparser==2.20
PyJWT==2.0.1
requests==2.25.1
setuptools==54.1.1
urllib3==1.26.3